refactor: extract app window components
